shouldn't seg fault on hats

+++ src/Hat.h
@@ -3,7 +3,7 @@
 
 class Hat {
  private:
-  int jump;
+  float jump;
   int dmgRed;
 
  public:
